Update README.md
Introduction

A project called Docusaurus makes it simple to create, distribute, and manage open source project sites. Use docusaurus.new to try Docusaurus in a playground right away.

Easy to Begin

Docusaurus is designed to launch into operation as quickly as feasible. To free you up to work on your project, we designed Docusaurus to manage website development.

Localizable

Through CrowdIn, Docusaurus has translation support. By interpreting your documents, you may strengthen and expand your global network.

Adjustable

Although Docusaurus comes pre-loaded with the essential pages and features you'll need to get going, such as a home page, a documents part, a weblog, and other help sites, it is also fully configurable so you can create a site that are all yours.

Implementation

The installation CLI should be used to build your site: npm init docusaurus@latest

For additional info, read the documents.



Making a contribution

We made Docusaurus available because it enables us to expand more effectively and supports the numerous OSS initiatives inside Facebook. We anticipate that the initiative will be useful to other organisations. Any donations from the communities are appreciated.

A code of behaviour

Members in the initiative are expected to abide by the Code of Conduct that Facebook has issued. Please read the entire document so that you are aware of the acceptable and unacceptable behaviour.

Assisting manual

To find out more about how we develop Docusaurus, how to suggest fixes and enhancements, and exactly how to create and test your modifications, visit our contribution instructions.

Easy-to-understand bugs

We have compiled a list of newcomer bugs that may contain simpler problems to address first in order to assist you get started and familiarise yourself with our contributing procedure. This is a fantastic starting point.

Contact

Several ways to reach us are available:
-	Use #general on Discord if you use Docusaurus.
-	#contributors is for anyone who wish to help the Docusaurus core.
-	Twitter user @docusaurus
-	GitHub Issues

Contributors

Thanks to everyone who has contributed, this project is possible.

Backers

To all of our supporters, thank you! Become a supporter.

Sponsors

Become a sponsor to aid our endeavour. Here, a connection to the website will be shown beside your logo. Join as a sponsor.

License

Docusaurus possesses an MIT licence.
The Docusaurus information is released under an Open Source license, such as the.md files in the /docs folder.

Extra gratitude

-	We are supported by BrowserStack's free open source accessibility.
-	We can detect Html tags or accessible problems with the aid of Rocket Validator.